Add array in fields + refactor common array and record tools
