+++ packages/ui/spa/ng/components/Module.tsx
@@ -1,35 +1,21 @@
-import { Internal, ModulePath, SourcePath } from "@valbuild/core";
-import {
-  useAddPatch,
-  useSchemaAtPath,
-  useShallowSourceAtPath,
-  useValPortal,
-} from "../ValProvider";
+import { SourcePath } from "@valbuild/core";
+import { useSchemaAtPath } from "../ValProvider";
 import { FieldSchemaError } from "./FieldSchemaError";
 import { FieldLoading } from "./FieldLoading";
 import { FieldNotFound } from "./FieldNotFound";
 import { AnyField } from "./AnyField";
-import { prettifyFilename } from "../../utils/prettifyFilename";
-import { Fragment, useEffect, useState } from "react";
-import { ChevronRight, Edit, Plus } from "lucide-react";
+import { Fragment } from "react";
+import { ChevronRight } from "lucide-react";
 import { useNavigation } from "../../components/ValRouter";
-import { array } from "@valbuild/core/fp";
-import { Input } from "../../components/ui/input";
-import { Button } from "../../components/ui/button";
-import { Popover, PopoverContent } from "../../components/ui/popover";
-import { PopoverTrigger } from "@radix-ui/react-popover";
-import { emptyOf } from "../../components/fields/emptyOf";
-import { JSONValue } from "@valbuild/core/patch";
+import {
+  ArrayAndRecordTools,
+  splitIntoInitAndLastParts,
+} from "./ArrayAndRecordTools";
+import { isParentArray, useParent } from "../hooks/useParent";
 
 export function Module({ path }: { path: SourcePath }) {
   const schemaAtPath = useSchemaAtPath(path);
-  const [moduleFilePath, modulePath] =
-    Internal.splitModuleFilePathAndModulePath(path);
-  const maybeParentPath = Internal.joinModuleFilePathAndModulePath(
-    moduleFilePath,
-    Internal.splitModulePath(modulePath).slice(0, -1).join(".") as ModulePath,
-  );
-  const parentSchemaAtPath = useSchemaAtPath(maybeParentPath);
+  const { path: maybeParentPath, schema: parentSchema } = useParent(path);
   const { navigate } = useNavigation();
   if (schemaAtPath.status === "loading") {
     return <FieldLoading path={path} type="module" />;
@@ -42,33 +28,12 @@ export function Module({ path }: { path: SourcePath }) {
   if (schemaAtPath.status === "not-found") {
     return <FieldNotFound path={path} type="module" />;
   }
-  if (parentSchemaAtPath.status === "loading") {
-    return <FieldLoading path={maybeParentPath} type="module" />;
-  }
-  if (parentSchemaAtPath.status === "error") {
-    return (
-      <FieldSchemaError
-        path={maybeParentPath}
-        error={parentSchemaAtPath.error}
-        type="module"
-      />
-    );
-  }
-  if (parentSchemaAtPath.status === "not-found") {
-    return <FieldNotFound path={maybeParentPath} type="module" />;
-  }
 
   const schema = schemaAtPath.data;
-  const isParentRecord =
-    maybeParentPath !== path && parentSchemaAtPath.data.type === "record";
-  const isParentArray =
-    maybeParentPath !== path && parentSchemaAtPath.data.type === "array";
-  const isRecord = schema.type === "record";
-  const isArray = schema.type === "array";
-
   const parts = splitIntoInitAndLastParts(path);
   const init = parts.slice(0, -1);
   const last = parts[parts.length - 1];
+  const showNumber = isParentArray(path, maybeParentPath, parentSchema);
   return (
     <div className="flex flex-col gap-6 py-10">
       <div className="flex flex-col gap-2 text-left">
@@ -106,303 +71,13 @@ export function Module({ path }: { path: SourcePath }) {
         )}
         <div>
           <div className="flex items-center justify-between h-12 gap-4 text-xl">
-            {!isParentArray && <span>{last.text}</span>}
-            {isParentArray && <span>#{Number(last.text) + 1}</span>}
-            <span className="inline-flex items-center gap-2">
-              {isParentRecord && (
-                <ChangeRecordPopover
-                  defaultValue={last.text}
-                  path={path}
-                  parentPath={maybeParentPath}
-                />
-              )}
-              {isArray && <AddArrayButton path={path} />}
-              {isRecord && <AddRecordPopover path={path} />}
-            </span>
+            {!showNumber && <span>{last.text}</span>}
+            {showNumber && <span>#{Number(last.text) + 1}</span>}
+            <ArrayAndRecordTools path={path} variant={"module"} />
           </div>
         </div>
       </div>
       <AnyField key={path} path={path} schema={schema} />
     </div>
   );
-}
-
-function AddArrayButton({ path }: { path: SourcePath }) {
-  const { navigate } = useNavigation();
-  const { addPatch, patchPath } = useAddPatch(path);
-  const schmeaAtPath = useSchemaAtPath(path);
-  const shallowSourceAtPath = useShallowSourceAtPath(path, "array");
-  const [moduleFilePath] = Internal.splitModuleFilePathAndModulePath(path);
-  if (!("data" in shallowSourceAtPath) || !shallowSourceAtPath.data) {
-    return null;
-  }
-  if (!("data" in schmeaAtPath)) {
-    return null;
-  }
-  const schema = schmeaAtPath.data;
-  if (schema.type !== "array") {
-    console.error("Cannot add to non-array", shallowSourceAtPath, {
-      parentPath: path,
-    });
-    return null;
-  }
-  const highestIndex = shallowSourceAtPath.data.length;
-  return (
-    <Button
-      title="Add"
-      onClick={() => {
-        const newPatchPath = patchPath.concat(highestIndex.toString());
-        console.log({ newPatchPath });
-        addPatch([
-          {
-            op: "add",
-            path: newPatchPath,
-            value: emptyOf(schema.item) as JSONValue,
-          },
-        ]);
-        navigate(
-          Internal.joinModuleFilePathAndModulePath(
-            moduleFilePath,
-            Internal.patchPathToModulePath(newPatchPath),
-          ),
-        );
-      }}
-    >
-      <Plus size={16} />
-    </Button>
-  );
-}
-
-function ChangeRecordPopover({
-  defaultValue,
-  path,
-  parentPath,
-}: {
-  defaultValue: string;
-  path: SourcePath;
-  parentPath: SourcePath;
-}) {
-  const [moduleFilePath, parentModulePath] =
-    Internal.splitModuleFilePathAndModulePath(parentPath);
-  const { navigate } = useNavigation();
-  const { addPatch } = useAddPatch(path);
-  const shallowParentSource = useShallowSourceAtPath(parentPath, "record");
-  const [key, setKey] = useState(defaultValue); // cannot change - right?
-  const [open, setOpen] = useState(false);
-  const portalContainer = useValPortal();
-  useEffect(() => {
-    const keyDownListener = (ev: KeyboardEvent) => {
-      if (ev.key === "Escape") {
-        setOpen(false);
-      }
-    };
-    window.addEventListener("keydown", keyDownListener);
-    return () => {
-      window.removeEventListener("keydown", keyDownListener);
-    };
-  }, []);
-
-  const parentPatchPath = Internal.createPatchPath(parentModulePath);
-  if (
-    !("data" in shallowParentSource) ||
-    shallowParentSource.data === undefined ||
-    shallowParentSource.data === null
-  ) {
-    // An actual error message should be shown above
-    console.error("Parent source not found", shallowParentSource, { path });
-    return null;
-  }
-  const disabled =
-    key === defaultValue || key === "" || key in shallowParentSource.data;
-  return (
-    <Popover open={open}>
-      <Button asChild>
-        <PopoverTrigger
-          onClick={() => {
-            setOpen(true);
-          }}
-          title="Change name"
-        >
-          <Edit size={16} />
-        </PopoverTrigger>
-      </Button>
-      <PopoverContent container={portalContainer} className="text-text-primary">
-        <form
-          className="flex flex-col gap-2"
-          onSubmit={(ev) => {
-            ev.preventDefault();
-            addPatch([
-              {
-                op: "move",
-                from: parentPatchPath.concat(
-                  defaultValue,
-                ) as array.NonEmptyArray<string>,
-                path: parentPatchPath.concat(
-                  key,
-                ) as array.NonEmptyArray<string>,
-              },
-            ]);
-            navigate(
-              Internal.joinModuleFilePathAndModulePath(
-                moduleFilePath,
-                Internal.patchPathToModulePath(parentPatchPath.concat(key)),
-              ),
-              {
-                replace: true,
-              },
-            );
-            setOpen(false);
-          }}
-        >
-          <Input
-            type="text"
-            value={key}
-            onChange={(ev) => {
-              setKey(ev.target.value);
-            }}
-          />
-          <div className="flex items-center gap-2">
-            <Button disabled={disabled} variant="outline" type="submit">
-              Update
-            </Button>
-
-            <Button
-              variant={"ghost"}
-              type="reset"
-              onClick={() => {
-                setOpen(false);
-                setKey(defaultValue);
-              }}
-            >
-              Cancel
-            </Button>
-          </div>
-        </form>
-      </PopoverContent>
-    </Popover>
-  );
-}
-
-function AddRecordPopover({ path }: { path: SourcePath }) {
-  const shallowSourceAtPath = useShallowSourceAtPath(path, "record");
-  const schemaAtPath = useSchemaAtPath(path);
-  const portalContainer = useValPortal();
-  const { addPatch } = useAddPatch(path);
-  const [key, setKey] = useState("");
-  const { navigate } = useNavigation();
-  const [open, setOpen] = useState(false);
-  useEffect(() => {
-    const keyDownListener = (ev: KeyboardEvent) => {
-      if (ev.key === "Escape") {
-        setOpen(false);
-      }
-    };
-    window.addEventListener("keydown", keyDownListener);
-    return () => {
-      window.removeEventListener("keydown", keyDownListener);
-    };
-  }, []);
-  if (
-    !("data" in shallowSourceAtPath) ||
-    shallowSourceAtPath.data === undefined ||
-    shallowSourceAtPath.data === null
-  ) {
-    // An actual error message should be shown above
-    console.error("Source not found", shallowSourceAtPath, { path });
-    return null;
-  }
-  if (!("data" in schemaAtPath) || schemaAtPath.data === undefined) {
-    // An actual error message should be shown above
-    console.error("Schema not found", shallowSourceAtPath, { path });
-    return null;
-  }
-  const schema = schemaAtPath.data;
-  if (schema.type !== "record") {
-    // An actual error message should be shown above
-    console.error("Schema is not a record", schemaAtPath, { path });
-    return null;
-  }
-  const disabled = key === "" || key in shallowSourceAtPath.data;
-  const [moduleFilePath, modulePath] =
-    Internal.splitModuleFilePathAndModulePath(path);
-  return (
-    <Popover open={open}>
-      <Button asChild>
-        <PopoverTrigger
-          onClick={() => {
-            setOpen(true);
-          }}
-          title={"Add"}
-        >
-          <Plus size={16} />
-        </PopoverTrigger>
-      </Button>
-      <PopoverContent container={portalContainer}>
-        <form
-          className="flex flex-col gap-2"
-          onSubmit={(ev) => {
-            ev.preventDefault();
-            const newPatchPath =
-              Internal.createPatchPath(modulePath).concat(key);
-            addPatch([
-              {
-                op: "add",
-                path: newPatchPath,
-                value: emptyOf(schema.item) as JSONValue,
-              },
-            ]);
-            navigate(
-              Internal.joinModuleFilePathAndModulePath(
-                moduleFilePath,
-                Internal.patchPathToModulePath(newPatchPath),
-              ) as SourcePath,
-            );
-            setOpen(false);
-          }}
-        >
-          <Input
-            autoFocus
-            value={key}
-            onChange={(ev) => {
-              setKey(ev.target.value);
-            }}
-          />
-          <div>
-            <Button type="submit" disabled={disabled} variant={"outline"}>
-              Add
-            </Button>
-          </div>
-        </form>
-      </PopoverContent>
-    </Popover>
-  );
-}
-
-function splitIntoInitAndLastParts(
-  path: SourcePath,
-): { text: string; sourcePath: SourcePath }[] {
-  const [moduleFilePath, modulePath] =
-    Internal.splitModuleFilePathAndModulePath(path);
-  const moduleFilePathParts = Internal.splitModuleFilePath(moduleFilePath).map(
-    (part) => {
-      return {
-        text: prettifyFilename(part),
-        sourcePath: moduleFilePath as unknown as SourcePath,
-      };
-    },
-  );
-  if (!modulePath) {
-    return moduleFilePathParts;
-  }
-  const splittedModulePath = Internal.splitModulePath(modulePath);
-  const modulePathParts = splittedModulePath.map((part, i) => {
-    return {
-      text: part,
-      sourcePath: Internal.joinModuleFilePathAndModulePath(
-        moduleFilePath,
-        splittedModulePath.slice(0, i + 1).join(".") as ModulePath,
-      ),
-    };
-  });
-  return moduleFilePathParts.concat(modulePathParts);
 }
